prefer type unknown to any
