+++ packages/core/src/types/type-info.ts
@@ -1,4 +1,6 @@
-import { type KeystoneContext } from './context'
+import {
+  type KeystoneContext
+} from './context'
 import { type BaseItem } from './next-fields'
 
 type GraphQLInput = Record<string, any>
